requirements.txt
flask>=3.0.0
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=15.0.0
ccxt>=4.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
